refactor: 판매자 & 일반 유저 API 분리 - Seller: 판매자 #2
