+++ src/main/java/poomasi/domain/farm/controller/FarmController.java
@@ -1,0 +1,20 @@
+package poomasi.domain.farm.controller;
+
+import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import poomasi.domain.farm.service.FarmService;
+
+
+@RestController()
+@RequestMapping("/api/farm")
+@AllArgsConstructor
+
+public class FarmController {
+    private final FarmService farmService;
+
+    @GetMapping("/{farmId}")
+    public ResponseEntity<?> getFarm(@RequestParam Long farmId) {
+        return ResponseEntity.ok(farmService.getFarmByFarmId(farmId));
+    }
+}
